Strategies/requirements.txt
gym==0.21.0
gym-anytrading==1.2.0
stable-baselines3==1.5.0
numpy
requests
pandas
python-dotenv
cloudpickle==2.0.0
